+++ modules/basemap-layers/src/util.ts
@@ -36,4 +36,3 @@ export function featuresArrayToObject(
 
   return {[sourceName]: featuresByLayer};
 }
-